requirements.txt
python-telegram-bot==20.8
pillow
groq
requests
python-dotenv
